Phase 2: collapse /api/v2 into canonical /api/redlist, remove dead CSV species path
The v2 routes were transitional dual-run scaffolding while the species read
layer moved onto DuckDB/Parquet. Now that it's the default, fold them back:

- /api/v2/species → /api/redlist/species (parquet-backed querySpecies); the old
  CSV-backed handler is replaced.
- /api/v2/species/history → /api/redlist/species/history.
- Delete /api/v2; point RedListView at the canonical routes.
- next.config: re-key the DuckDB trace includes + data/** excludes to the
  canonical routes; drop the v2 entries.

Remove the now-dead v1 CSV species path:
- species-store.getSpecies + its GBIF/mapping-only helpers (loadMapping,
  loadGbifForGroup, parseMappingRow, parseGbifRow), the gbif/mapping caches,
  and the now-unused GbifRow/MappingRow/MappingLink/SpeciesRow types + their
  tests. loadRedlistForGroup/loadHistoryForGroup stay (assessor routes use them).
- Strip the unused useRedListSpecies hook to just the RedListSpecies type
  (still imported by RedListView/exportCsv).

Pure refactor — no data change, no re-sync. 348 tests pass, tsc + lint clean.
