+++ app/src/app/api/redlist/species/history/route.ts
@@ -1,6 +1,6 @@
 /**
- * v2 lazy assessment history (#260 Phase 2). The species list no longer carries
- * the full per-species history array (it's ~half the payload); the detail panel
+ * Lazy assessment history (#260 Phase 2). The species list no longer carries the
+ * full per-species history array (it's ~half the payload); the detail panel
  * fetches it here on open. ?id=<sisTaxonId> → { previous_assessments }.
  */
 import { NextRequest, NextResponse } from "next/server";
@@ -16,8 +16,8 @@ export async function GET(request: NextRequest) {
     const previous_assessments = await getAssessmentHistory(id);
     return NextResponse.json({ previous_assessments }, { headers: CACHE_5M });
   } catch (error) {
-    console.error("v2 history query error:", error);
+    console.error("history query error:", error);
     const message = error instanceof Error ? error.message : "Unknown error";
-    return NextResponse.json({ error: `v2 history query failed: ${message}` }, { status: 500 });
+    return NextResponse.json({ error: `history query failed: ${message}` }, { status: 500 });
   }
 }
