+++ app/src/app/api/redlist/species/route.ts
@@ -1,32 +1,27 @@
+/**
+ * Species list — DuckDB/Parquet-backed (querying the parquets in R2), plus
+ * arbitrary-rank filtering (e.g. ?taxon=turdidae). Same params + response as
+ * before; the full per-species assessment history is fetched lazily via the
+ * /api/redlist/species/history sub-route.
+ */
 import { NextRequest, NextResponse } from "next/server";
-import { getSpecies } from "@/lib/data/species-store";
-import { getCsvGroupsForNode } from "@/lib/taxonomy-utils";
+import { querySpecies } from "@/lib/data/species-duckdb";
 import { CACHE_5M } from "@/lib/cache-headers";
 
 export async function GET(request: NextRequest) {
-  const searchParams = request.nextUrl.searchParams;
-  const taxonId = searchParams.get("taxon") || "all";
-  const category = searchParams.get("category");
-  const groups = getCsvGroupsForNode(taxonId);
+  const sp = request.nextUrl.searchParams;
+  const taxon = sp.get("taxon") || "all";
+  const category = sp.get("category");
 
   try {
     const includeNE = category === "NE";
-    let species = getSpecies(groups, includeNE);
+    let species = await querySpecies({ taxon, includeNE });
+    if (category === "NE") species = species.filter((s) => s.category === "NE");
 
-    if (category === "NE") {
-      species = species.filter((s) => s.category === "NE");
-    }
-
-    return NextResponse.json(
-      { species, total: species.length },
-      { headers: CACHE_5M }
-    );
+    return NextResponse.json({ species, total: species.length }, { headers: CACHE_5M });
   } catch (error) {
-    console.error("Species query error:", error);
+    console.error("species query error:", error);
     const message = error instanceof Error ? error.message : "Unknown error";
-    return NextResponse.json(
-      { error: `Species query failed: ${message}` },
-      { status: 500 }
-    );
+    return NextResponse.json({ error: `species query failed: ${message}` }, { status: 500 });
   }
 }
