+++ app/src/lib/data/species-store.test.ts
@@ -12,7 +12,7 @@ vi.mock("./csv", () => ({
 
 import * as fs from "fs";
 import { readCsv } from "./csv";
-import { getAssessorCandidates, getAssessorCandidatesByCountry, getSpecies, _resetCaches } from "./species-store";
+import { getAssessorCandidates, getAssessorCandidatesByCountry } from "./species-store";
 
 // ---------------------------------------------------------------------------
 // Helpers
@@ -638,252 +638,3 @@ describe("getAssessorCandidatesByCountry", () => {
     expect(beetleResult[0].name).toBe("Beetle Expert");
   });
 });
-
-// ---------------------------------------------------------------------------
-// getSpecies — 1:N mapping (species → multiple GBIF keys via synonyms)
-// ---------------------------------------------------------------------------
-
-describe("getSpecies (1:N mapping)", () => {
-  // The mapping CSV has columns: sis_taxon_id, gbif_species_key, match_type, name_source.
-  // GBIF CSV is parsed by parseGbifRow with columns:
-  //   gbif_species_key, scientific_name, common_name, taxon_group_table1a,
-  //   total_count, count_after_assessment_year, class_name, order_name, family, countries
-  // Redlist CSV uses parseRedlistRow.
-
-  interface MappingFixture {
-    sis_taxon_id: number;
-    gbif_species_key: number | null;
-    match_type: string;
-    name_source?: string;
-  }
-  interface GbifFixture {
-    gbif_species_key: number;
-    scientific_name: string;
-    total_count: number;
-    count_after_assessment_year: number | null;
-    common_name?: string;
-    family?: string;
-    class_name?: string;
-    order_name?: string;
-  }
-
-  /**
-   * Set up readCsv to dispatch by path: mapping.csv → mappingRows,
-   * redlist/<group>.csv → redlistRows, gbif/<group>.csv → gbifRows.
-   */
-  function setupGetSpecies(opts: {
-    group: string;
-    redlistRows?: TestRedlistRow[];
-    mappingRows?: MappingFixture[];
-    gbifRows?: GbifFixture[];
-  }) {
-    _resetCaches();
-    const redlist = opts.redlistRows ?? [];
-    const mapping = (opts.mappingRows ?? []).map((m) => ({
-      ...m,
-      name_source: m.name_source ?? "",
-    }));
-    const gbif = (opts.gbifRows ?? []).map((g) => ({
-      gbif_species_key: g.gbif_species_key,
-      scientific_name: g.scientific_name,
-      common_name: g.common_name ?? "",
-      taxon_group_table1a: opts.group,
-      total_count: g.total_count,
-      count_after_assessment_year: g.count_after_assessment_year,
-      class_name: g.class_name ?? "",
-      order_name: g.order_name ?? "",
-      family: g.family ?? "",
-      countries: [] as string[],
-    }));
-
-    vi.mocked(fs.existsSync).mockReturnValue(true);
-    vi.mocked(fs.readFileSync).mockReturnValue("{}"); // empty history
-    vi.mocked(readCsv).mockImplementation(((csvPath: string) => {
-      const p = String(csvPath);
-      if (p.endsWith("mapping.csv")) return mapping as ReturnType<typeof readCsv>;
-      if (p.includes("/gbif/")) return gbif as ReturnType<typeof readCsv>;
-      // Anything else (per-group redlist) → redlist rows.
-      return redlist as ReturnType<typeof readCsv>;
-    }) as typeof readCsv);
-  }
-
-  it("regression: single GBIF link aggregates correctly", () => {
-    const group = uniqueGroup();
-    const row = makeRow({ scientific_name: "Aquarana catesbeianus", taxon_group_table1a: group });
-    setupGetSpecies({
-      group,
-      redlistRows: [row],
-      mappingRows: [
-        { sis_taxon_id: row.sis_taxon_id, gbif_species_key: 100, match_type: "EXACT", name_source: "canonical" },
-      ],
-      gbifRows: [
-        { gbif_species_key: 100, scientific_name: "Aquarana catesbeianus", total_count: 50, count_after_assessment_year: 10 },
-      ],
-    });
-
-    const results = getSpecies([group], false);
-    expect(results).toHaveLength(1);
-    expect(results[0].gbif_species_key).toBe(100);
-    expect(results[0].gbif_occurrence_count).toBe(50);
-    expect(results[0].gbif_observations_after_assessment_year).toBe(10);
-  });
-
-  it("multiple GBIF links sum occurrence counts", () => {
-    const group = uniqueGroup();
-    const row = makeRow({ scientific_name: "Aquarana catesbeianus", taxon_group_table1a: group });
-    setupGetSpecies({
-      group,
-      redlistRows: [row],
-      mappingRows: [
-        { sis_taxon_id: row.sis_taxon_id, gbif_species_key: 100, match_type: "EXACT", name_source: "canonical" },
-        { sis_taxon_id: row.sis_taxon_id, gbif_species_key: 200, match_type: "EXACT", name_source: "synonym" },
-      ],
-      gbifRows: [
-        { gbif_species_key: 100, scientific_name: "Aquarana catesbeianus", total_count: 50, count_after_assessment_year: 10 },
-        { gbif_species_key: 200, scientific_name: "Lithobates catesbeianus", total_count: 30, count_after_assessment_year: 5 },
-      ],
-    });
-
-    const results = getSpecies([group], false);
-    expect(results).toHaveLength(1);
-    expect(results[0].gbif_occurrence_count).toBe(80);
-    expect(results[0].gbif_observations_after_assessment_year).toBe(15);
-    // Primary (displayed) key is the first non-null link → canonical (100).
-    expect(results[0].gbif_species_key).toBe(100);
-  });
-
-  it("multiple GBIF links populate linkedGbifKeys (both excluded from NE list)", () => {
-    const group = uniqueGroup();
-    const row = makeRow({ scientific_name: "Aquarana catesbeianus", taxon_group_table1a: group });
-    setupGetSpecies({
-      group,
-      redlistRows: [row],
-      mappingRows: [
-        { sis_taxon_id: row.sis_taxon_id, gbif_species_key: 100, match_type: "EXACT" },
-        { sis_taxon_id: row.sis_taxon_id, gbif_species_key: 200, match_type: "EXACT" },
-      ],
-      gbifRows: [
-        { gbif_species_key: 100, scientific_name: "Aquarana catesbeianus", total_count: 50, count_after_assessment_year: 10 },
-        { gbif_species_key: 200, scientific_name: "Lithobates catesbeianus", total_count: 30, count_after_assessment_year: 5 },
-        { gbif_species_key: 300, scientific_name: "Other species", total_count: 5, count_after_assessment_year: 1 },
-      ],
-    });
-
-    const results = getSpecies([group], true);
-    // 1 redlist row + 1 NE (key 300 only). Keys 100 and 200 are linked.
-    const ne = results.filter((r) => r.category === "NE");
-    expect(ne).toHaveLength(1);
-    expect(ne[0].gbif_species_key).toBe(300);
-  });
-
-  it("missing GBIF data on one of multiple keys: only available counts are summed", () => {
-    const group = uniqueGroup();
-    const row = makeRow({ scientific_name: "Aquarana catesbeianus", taxon_group_table1a: group });
-    setupGetSpecies({
-      group,
-      redlistRows: [row],
-      mappingRows: [
-        { sis_taxon_id: row.sis_taxon_id, gbif_species_key: 100, match_type: "EXACT" },
-        { sis_taxon_id: row.sis_taxon_id, gbif_species_key: 999, match_type: "EXACT" }, // not in gbifMap
-      ],
-      gbifRows: [
-        { gbif_species_key: 100, scientific_name: "Aquarana catesbeianus", total_count: 50, count_after_assessment_year: 10 },
-      ],
-    });
-
-    const results = getSpecies([group], false);
-    expect(results).toHaveLength(1);
-    expect(results[0].gbif_occurrence_count).toBe(50);
-    expect(results[0].gbif_observations_after_assessment_year).toBe(10);
-    expect(Number.isNaN(results[0].gbif_occurrence_count!)).toBe(false);
-  });
-
-  it("primary gbif_species_key prefers canonical regardless of mapping CSV row order", () => {
-    // The matcher writes canonical rows before synonym rows today, but the
-    // reader must not depend on that — selection should be by name_source.
-    // This fixture deliberately lists the synonym row FIRST.
-    const group = uniqueGroup();
-    const row = makeRow({ scientific_name: "Aquarana catesbeianus", taxon_group_table1a: group });
-    setupGetSpecies({
-      group,
-      redlistRows: [row],
-      mappingRows: [
-        { sis_taxon_id: row.sis_taxon_id, gbif_species_key: 200, match_type: "EXACT", name_source: "synonym" },
-        { sis_taxon_id: row.sis_taxon_id, gbif_species_key: 100, match_type: "EXACT", name_source: "canonical" },
-      ],
-      gbifRows: [
-        { gbif_species_key: 100, scientific_name: "Aquarana catesbeianus", total_count: 50, count_after_assessment_year: 10 },
-        { gbif_species_key: 200, scientific_name: "Lithobates catesbeianus", total_count: 30, count_after_assessment_year: 5 },
-      ],
-    });
-
-    const results = getSpecies([group], false);
-    expect(results).toHaveLength(1);
-    // Despite the synonym row appearing first in the mapping, the canonical
-    // key (100) is selected as the displayed/external-link gbif_species_key.
-    expect(results[0].gbif_species_key).toBe(100);
-    // Counts still aggregate across both keys.
-    expect(results[0].gbif_occurrence_count).toBe(80);
-  });
-
-  it("falls back to first non-null link when no canonical-source link exists", () => {
-    // Edge case: a species whose canonical name didn't match GBIF at all,
-    // but a synonym did. The displayed gbif_species_key should be the
-    // synonym key (only option), not null.
-    const group = uniqueGroup();
-    const row = makeRow({ scientific_name: "Aquarana catesbeianus", taxon_group_table1a: group });
-    setupGetSpecies({
-      group,
-      redlistRows: [row],
-      mappingRows: [
-        { sis_taxon_id: row.sis_taxon_id, gbif_species_key: 200, match_type: "EXACT", name_source: "synonym" },
-      ],
-      gbifRows: [
-        { gbif_species_key: 200, scientific_name: "Lithobates catesbeianus", total_count: 30, count_after_assessment_year: 5 },
-      ],
-    });
-
-    const results = getSpecies([group], false);
-    expect(results[0].gbif_species_key).toBe(200);
-    expect(results[0].gbif_occurrence_count).toBe(30);
-  });
-
-  it("Bullfrog regression: legacy Lithobates key is linked to Aquarana, not surfaced as NE", () => {
-    // The user-reported bug: Lithobates catesbeianus appearing as a separate
-    // NE row even though Aquarana catesbeianus has an assessment. With the
-    // 1:N mapping, both GBIF keys link back to the bullfrog assessment.
-    const group = uniqueGroup();
-    const bullfrog = makeRow({
-      sis_taxon_id: 58565,
-      scientific_name: "Aquarana catesbeianus",
-      common_name: "American Bullfrog",
-      category: "LC",
-      taxon_group_table1a: group,
-    });
-    setupGetSpecies({
-      group,
-      redlistRows: [bullfrog],
-      mappingRows: [
-        { sis_taxon_id: 58565, gbif_species_key: 5217419, match_type: "EXACT", name_source: "canonical" },
-        { sis_taxon_id: 58565, gbif_species_key: 2427091, match_type: "EXACT", name_source: "synonym" },
-      ],
-      gbifRows: [
-        { gbif_species_key: 5217419, scientific_name: "Aquarana catesbeianus", total_count: 100, count_after_assessment_year: 20 },
-        { gbif_species_key: 2427091, scientific_name: "Lithobates catesbeianus", total_count: 5000, count_after_assessment_year: 800 },
-      ],
-    });
-
-    const results = getSpecies([group], true);
-    // Exactly one row, no NE row for the legacy Lithobates key.
-    expect(results).toHaveLength(1);
-    const bullfrogResult = results[0];
-    expect(bullfrogResult.scientific_name).toBe("Aquarana catesbeianus");
-    expect(bullfrogResult.category).toBe("LC");
-    // All occurrences (5100) flow to the assessed species.
-    expect(bullfrogResult.gbif_occurrence_count).toBe(5100);
-    expect(bullfrogResult.gbif_observations_after_assessment_year).toBe(820);
-    // No "Lithobates catesbeianus" NE row.
-    expect(results.find((r) => r.scientific_name === "Lithobates catesbeianus")).toBeUndefined();
-  });
-});
-
